Modify sample, fix driver bug
Driver bug fixed where new sensitivity is not set if its lower
then the previous one. Example with the bug, if the motor is running
with 50% duty cycle and now the motor have to set to 30%, the motor
would still run with 50%.

Extend the sample for 10 frames and 100 frames mode. The sample starts
with a 0% sensitivity for 8 motors and increment by 10% each 5 seconds.
If the sensitivity is > 100 it starts with 0 again.