drivers: sensor: bno055: add bosch bno055 sensor
Add support for the bosch sensor bno055. The BNO055 is a System in
Package (SiP), integrating a triaxial 14-bit accelerometer, a triaxial
16-bit gyroscope with a range of ±2000 degrees per second, a triaxial
geomagnetic sensor and a 32-bit cortex M0+ microcontroller running
Bosch Sensortec sensor fusion software, in a single package.

The driver supports in the initial commit only a small part of the
sensor. Reading the euler angles, running the sensor in the fusion modes
and use the i2c-bus.

Signed-off-by: Tobias Mierzwa <tobias.mierzwa@web.de>